package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remote-behave-steps"
version = "0.1.0"
description = "Distributed BDD testing with Python Behave via remote step providers"
requires-python = ">=3.10"
dependencies = [
    "behave>=1.2.6",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flask>=3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/remote_behave_steps"]

[tool.pytest.ini_options]
testpaths = ["tests"]

# Example configuration for remote_behave_steps consumers:
# [tool.remote_behave_steps]
# cache_ttl = 300
#
# [[tool.remote_behave_steps.servers]]
# name = "todo-service"
# url = "http://localhost:8080/openapi.yaml"
